+++ junseoungbin_new/src/main/java/com/likelion/junseoungbin_new/member/domain/Part.java
@@ -1,0 +1,7 @@
+package com.likelion.junseoungbin_new.member.domain;
+
+public enum Part {
+    Leader,
+    Follower,
+    Watcher
+}
